% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviations.R
\name{addAbbreviations}
\alias{addAbbreviations}
\title{Add abbreviations}
\usage{
addAbbreviations(termDetector, token, abbreviation)
}
\arguments{
\item{termDetector}{:a TermDetector instance}

\item{token}{: the token (ex : "ac" in "acide acetique")}

\item{abbreviation}{:the abbreviation (ex: "acide" for the tokne "ac")}
}
\value{
Nothing, it just add the abbreviation
}
\description{
Add abbreviation at the token level to help the detector
}
\examples{
termDetector <- RIAMsystem::newTermDetector()
}
