% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTerms.R
\name{addTerms}
\alias{addTerms}
\title{Add terms to a TermDetector instance}
\usage{
addTerms(termDetector, terms, codes = NULL)
}
\arguments{
\item{termDetector}{:a TermDetector instance}

\item{terms}{:a character vector}

\item{codes}{:a character vector of the same length as terms. Default to NULL ("NA")}
}
\description{
Add terms and codes to a TermDetector instance to detect these in textual content. Use loadTerminoCSV() to load a whole terminology file.
}
\examples{
termDetector <- RIAMsystem::newTermDetector()
}
