% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermDetector.R
\name{detect}
\alias{detect}
\title{Detect terms in textual content}
\usage{
detect(termDetector, text)
}
\arguments{
\item{termDetector}{:a TermDetector instance}

\item{text}{:a character vector of length 1}
}
\value{
A dataframe containing:
}
\description{
Detect all the terms in textual content. Load the terms in the detector first with addTerms or loadTerminoCSV. It returns a data.frame with terms detected.
}
\examples{
termDetector <- RIAMsystem::newTermDetector()
}
