% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTerms.R
\name{loadTerminoCSV}
\alias{loadTerminoCSV}
\title{Load a CSV file of a terminology}
\usage{
loadTerminoCSV(termDetector, file, sep, colLabel, colCode)
}
\arguments{
\item{termDetector}{:a TermDetector instance}

\item{file}{:the name of the file which the data are to be read from (absolute or relative)}

\item{sep}{:the field separator character}

\item{colLabel}{: the ith column contaning the labels. Java uses zero-based indexing, first column is 0.}

\item{colCode}{: the ith column contaning the codes. Java uses zero-based indexing, first column is 0.}
}
\value{
Nothing, it just load the file
}
\description{
Load in Java a text file containing all the terms of a terminology
}
\examples{
termDetector <- RIAMsystem::newTermDetector()
}
