% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romedi.R
\docType{data}
\name{romedi}
\alias{romedi}
\title{Terminology example}
\format{An object of class \code{data.frame} with 10594 rows and 4 columns.}
\source{
\url{https://www.romedi.fr}
}
\usage{
romedi
}
\description{
Drug brand name and ingredients in France
}
\details{
version Romedi2.0.0
}
\author{
\describe{
  \item{instance}{codes}
  \item{type}{BN for BrandName, IN for ingredient}
  \item{label}{the label}
  \item{typeLabel}{SKOS: prefLabel, altLabel or HiddenLabel}
}
}
\keyword{datasets}
